move core component file
