+++ src/components/Core.ts
