+++ src/components/Header.ts
@@ -1,4 +1,4 @@
-import { Core } from '../util/Core';
+import { Core } from './Core';
 import { Navigation } from './Navigation';
 import { APP_TITLE_INNTERHTML, HEADER_INNERHTML } from '../constants/innerHTML';
